requirements.txt
requests>=2.31.0
Pillow>=10.0.0
# Optional – uncomment for PDF export:
# reportlab>=4.0.0
